Delete stuff from unmock-core.
