+++ packages/unmock-core/src/index.ts
@@ -2,10 +2,8 @@
 import * as fetch from "node-fetch";
 import * as sinon from "sinon";
 import Backend from "./backend";
-import NodeBackend from "./backend/node";
 import { ILogger, IUnmockOptions, IUnmockPackage } from "./interfaces";
-import FsSnapshotter, { ISnapshot } from "./loggers/snapshotter";
-import WinstonLogger from "./loggers/winston-logger";
+import { ISnapshot } from "./interfaces";
 import { ExtendedJSONSchema, nockify, vanillaJSONSchemify } from "./nock";
 import internalRunner, { IRunnerOptions } from "./runner";
 import { AllowedHosts, BooleanSetting } from "./settings";
@@ -17,8 +15,6 @@ export { u } from "./nock";
 export { transform, Addl, Arr } from "./generator-utils";
 
 export { ISnapshot };
-const utils = { snapshotter: FsSnapshotter };
-export { utils };
 
 export { Backend };
 
@@ -114,12 +110,3 @@ export class UnmockPackage implements IUnmockPackage {
     Object.values(this.backend.services).forEach(service => service.reset());
   }
 }
-
-export const unmock = new UnmockPackage(new NodeBackend(), {
-  logger: new WinstonLogger(),
-});
-
-export const nock = unmock.nock.bind(unmock);
-export const runner = unmock.runner.bind(unmock);
-
-export default unmock;
