+++ packages/unmock-core/src/interceptor/index.ts
@@ -1,5 +1,4 @@
 import { ISerializedRequest, ISerializedResponse } from "../interfaces";
-import NodeInterceptor from "./node-interceptor";
 
 /**
  * Interceptor callback. Called after request is serialized with
@@ -30,12 +29,4 @@ export type IInterceptorConstructor = new (
  */
 export interface IInterceptor {
   disable(): void;
-}
-
-/**
- * Create an interceptor and start intercepting requests.
- * @param options
- */
-export function createInterceptor(options: IInterceptorOptions): IInterceptor {
-  return new NodeInterceptor(options);
 }
