--- packages/unmock-core/src/loggers/winston-logger.ts
@@ -1,27 +1,0 @@
-import * as winston from "winston";
-import { ILogger } from "../interfaces";
-
-const logger = winston.createLogger({
-  levels: { unmock: 2 },
-  transports: [
-    new winston.transports.Console({
-      format: winston.format.combine(
-        winston.format.colorize(),
-        winston.format.simple(),
-      ),
-      level: "unmock",
-      stderrLevels: ["unmock"],
-    }),
-  ],
-});
-
-winston.addColors({ unmock: "cyan bold" });
-
-export default class WinstonLogger implements ILogger {
-  public log(message: string) {
-    logger.log({
-      level: "unmock",
-      message,
-    });
-  }
-}
