+++ packages/unmock-node/src/index.ts
@@ -1,6 +1,10 @@
 import { UnmockPackage } from "unmock-core";
 import NodeBackend from "./backend";
+import FsSnapshotter from "./loggers/snapshotter";
 import WinstonLogger from "./loggers/winston-logger";
+
+const utils = { snapshotter: FsSnapshotter };
+export { utils };
 
 export const unmock = new UnmockPackage(new NodeBackend(), {
   logger: new WinstonLogger(),
