+++ packages/unmock/src/index.ts
@@ -1,4 +1,4 @@
-import unmock from "unmock-core";
-
 export * from "unmock-core";
+import unmock from "unmock-node";
+export * from "unmock-node";
 export default unmock;
